reworked secondary function for HullModData
